refactor: extract pool selection policy
